@@ -7,12 +7,23 @@ using System;
 using System.IO;
 using System.Collections.Generic;
 using System.ComponentModel;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using Newtonsoft.Json;
-using Newtonsoft.Json.Converters;
-using Moesif.Api;
+//using System.Linq;
+//using System.Text;
+//using System.Threading.Tasks;
+//using Newtonsoft.Json;
+//using Newtonsoft.Json.Converters;
+//using Moesif.Api;
+
+#if NET6_0_OR_GREATER
+    using System.Text.Json.Serialization;
+#else
+     using System.Linq;
+     using System.Text;
+     using System.Threading.Tasks;
+     using Newtonsoft.Json;
+     using Newtonsoft.Json.Converters;
+//     using System.Text.Json;
+#endif
 
 namespace Moesif.Api.Models
 {
@@ -31,7 +42,11 @@ namespace Moesif.Api.Models
         /// <summary>
         /// Time when request was made
         /// </summary>
+#if NET6_0_OR_GREATER
+        [JsonPropertyName("time")]
+#else
         [JsonProperty("time")]
+#endif
         public DateTime Time 
         { 
             get 
@@ -48,7 +63,11 @@ namespace Moesif.Api.Models
         /// <summary>
         /// full uri of request such as https://www.example.com/my_path?param=1
         /// </summary>
+#if NET6_0_OR_GREATER
+        [JsonPropertyName("uri")]
+#else
         [JsonProperty("uri")]
+#endif
         public string Uri 
         { 
             get 
@@ -65,7 +84,11 @@ namespace Moesif.Api.Models
         /// <summary>
         /// verb of the API request such as GET or POST
         /// </summary>
+#if NET6_0_OR_GREATER
+        [JsonPropertyName("verb")]
+#else
         [JsonProperty("verb")]
+#endif
         public string Verb 
         { 
             get 
@@ -82,7 +105,11 @@ namespace Moesif.Api.Models
         /// <summary>
         /// Key/Value map of request headers
         /// </summary>
+#if NET6_0_OR_GREATER
+        [JsonPropertyName("headers")]
+#else
         [JsonProperty("headers")]
+#endif
         public Dictionary<string, string> Headers 
         { 
             get 
@@ -99,7 +126,11 @@ namespace Moesif.Api.Models
         /// <summary>
         /// Optionally tag the call with your API or App version
         /// </summary>
+#if NET6_0_OR_GREATER
+        [JsonPropertyName("api_version")]
+#else
         [JsonProperty("api_version")]
+#endif
         public string ApiVersion 
         { 
             get 
@@ -116,7 +147,11 @@ namespace Moesif.Api.Models
         /// <summary>
         /// IP Address of the client if known.
         /// </summary>
+#if NET6_0_OR_GREATER
+        [JsonPropertyName("ip_address")]
+#else
         [JsonProperty("ip_address")]
+#endif
         public string IpAddress 
         { 
             get 
@@ -133,7 +168,11 @@ namespace Moesif.Api.Models
         /// <summary>
         /// Request body
         /// </summary>
+#if NET6_0_OR_GREATER
+        [JsonPropertyName("body")]
+#else
         [JsonProperty("body")]
+#endif
         public object Body 
         { 
             get 
@@ -150,7 +189,11 @@ namespace Moesif.Api.Models
         /// <summary>
         /// Transfer Encoding of the body such as "base64", null value implies "json" transfer encoding.
         /// </summary>
+#if NET6_0_OR_GREATER
+        [JsonPropertyName("transfer_encoding")]
+#else
         [JsonProperty("transfer_encoding")]
+#endif
         public string TransferEncoding
         {
             get
